refactor: move `math` files to `solvers`
This is to avoid spending too much time organizing packages and
visibility until more structure is needed.
