+++ internal/solvers/algo.go
@@ -15,7 +15,7 @@
  along with this program.  If not, see <https://www.gnu.org/licenses/>.
 */
 
-package math
+package solvers
 
 import "log"
