+++ internal/solvers/algo_test.go
@@ -14,7 +14,7 @@ GNU Affero General Public License for more details.
 You should have received a copy of the GNU Affero General Public License
 along with this program.  If not, see <https://www.gnu.org/licenses/>.
 */
-package math
+package solvers
 
 import (
 	"testing"
